Cloud-Computing/api-model-py/requirements.txt
Flask==3.0.3
tensorflow==2.18.0
pandas==2.2.3
requests==2.32.3
Pillow==10.4.0
openpyxl==3.1.5
gunicorn==23.0.0
